Early menu/state support
This has a working state stack, meaning that menus and the ability to go
to other screens besides the main game are on their way! With this build
is a 'demo' menu that will pause all sounds upon entering and will
resume them on close. Use right click to acess the menu and left click
to leave it. Will be updating this soon with the full ability to
customize states and have user defined ones.