Fixed Menu
The menu is now customizable, with commands in the calls.txt, allowing
the user to display images, that can act as buttons if desired, as well
as the restart "r" button clears all assets and does a clean restart for
quick testing. The delay at the begining of the game is now removed so
the first user click starts the text instantly!